add: request and response all active rule IDs add: request and response rule data fix: no ctrl+c when waiting for wrapper's shared memory fix: conversation id logic for shm fix: terminal ui logs were cutoff
fuck: the conversation ID data types are cooked - lets hope they dont send more than 2 billion conversations thats UB

+++ scout/scout_wrapper/src/models/rules/response.rs
@@ -1,7 +1,9 @@
 use crate::constants::MAX_RESPONSES;
+use bytemuck::{Pod, Zeroable};
 
-#[repr(C)]
-#[derive(Clone, Copy)]
+// NOTE: repr(u32) to match C enum size (int = 4 bytes)
+#[repr(u32)]
+#[derive(Clone, Copy, PartialEq, Eq)]
 pub enum ResponseType {
     AirGap = 0,
     Kill = 1,
@@ -31,7 +33,8 @@ impl Response {
 }
 
 #[repr(C)]
+#[derive(Clone, Copy)]
 pub struct ResponseList {
     pub responses: [Response; MAX_RESPONSES],
-    pub length: usize,
+    pub length: u32,
 }
